调试框架覆盖率整合报告部分逻辑1
[Testframe] 报告整合提示出错，修改引入
test/tools/br/report.php : ln - 99, include
